feat: Add "=" at end of history label
When the user hits an operation button, an = is added on the end of the
UILabel. Thus the user will be able to tell whether the number in the
Calculator’s display is the result of a calculation or a number that the
user has just entered. Avoid multiple occurrences of = in the UILabel.

Resolves: Extra Credit #2