feat: Add ᐩ/- operation
Changes the sign of the number in the display. If the user is in the
middle of entering a number, changes the sign of that number and
allows typing to continue. On the other hand, if the user is not in
the middle of typing a number, then operation works like other
unary operations.

Resolves: Extra Credit #3